Add support for shellv2 style command
In order to get more meaningful output from adb commands
like the status code, etc. Support running adb shell
commands as shell,v2.

Test: need to be written
Bug: 74068064

+++ tests/src/com/android/tradefed/device/TestDeviceFuncTest.java
@@ -515,7 +515,7 @@ public class TestDeviceFuncTest implements IDeviceTest {
             Log.i(LOG_TAG, "Fastboot not enabled skipping testExecuteFastbootCommand_deviceInAdb");
             return;
         }
-        int origTimeout = mTestDevice.getCommandTimeout();
+        long origTimeout = mTestDevice.getCommandTimeout();
         try {
             assertEquals(TestDeviceState.ONLINE, mMonitor.getDeviceState());
             // reset operation timeout to small value to make test run quicker
